Add fields to subnet and nodein L2agent API spec
The schema `Subnet` now has 2 additional properties (address, and
mask) which is are essential information about a given subnet.

The schema `Node` no longer  has the field `subnet` and it's replaced by
`bia` which is the mac address assigned to the network device by the
manufacturer, because a node has multiple LTPs that each can be assigned
to a subnet rather than assigned to the device itself.